% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{add_line_break}
\alias{add_line_break}
\title{helpers}
\usage{
add_line_break(x = "very many many characters and words and sentences", n = 20)
}
\value{
The return value, if any, from executing the utility.

add_line_break

string
}
\description{
A utils function

A add a line break
}
