% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{add_line_break2}
\alias{add_line_break2}
\title{add_line_break2}
\usage{
add_line_break2(
  x = "very many many characters and words and sentences",
  n = 20
)
}
\value{
string
}
\description{
A add a line break2
}
