% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indicator.R
\name{alt_txt_indicator}
\alias{alt_txt_indicator}
\title{Alt text function}
\usage{
alt_txt_indicator(
  which_plot = "dotplot",
  input_value_regio_show_mode = "kaikki tason alueet",
  input_value_variable = "Muuttujan nimi",
  input_value_regio_level = "Hyvinvointialueet",
  input_value_region_selected = "Etelä-Karjan HVA"
)
}
\arguments{
\item{which_plot}{A string.}

\item{input_value_regio_show_mode}{A string.}

\item{input_value_variable}{A string.}

\item{input_value_regio_level}{A string.}

\item{input_value_region_selected}{A string.}
}
\description{
Alt text function
}
