% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_zipcode.R
\name{alt_txt_zipcode}
\alias{alt_txt_zipcode}
\title{Alt text function}
\usage{
alt_txt_zipcode(
  which_plot = "dotplot",
  input_value_variable = "Huono-osaisuus yhteensä",
  input_value_regio_level = "Hyvinvointialueet",
  input_value_region_selected = "Etelä-Karjalan HVA"
)
}
\arguments{
\item{which_plot}{A string.}

\item{input_value_variable}{A string.}

\item{input_value_regio_level}{A string.}

\item{input_value_region_selected}{A string.}
}
\description{
Alt text function
}
