% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alueprofiili.R
\name{create_alt_text_kartta}
\alias{create_alt_text_kartta}
\title{Process data for profile}
\usage{
create_alt_text_kartta(
  varname = NULL,
  input_value_regio_level_profile = "Hyvinvointialueet",
  input_value_region_profile = "Itä-Uudenmaan HVA"
)
}
\arguments{
\item{varname}{A string.}
}
\description{
Process data for profile
}
