% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alueprofiili.R
\name{create_alueprofiili_content}
\alias{create_alueprofiili_content}
\title{Process data for profile}
\usage{
create_alueprofiili_content(
  input_value_region_profile = "Itä-Uudenmaan HVA",
  input_value_regio_level_profile = "Hyvinvointialueet",
  type = "html",
  aikasarja = FALSE
)
}
\arguments{
\item{input_value_region_profile}{A string.}

\item{input_value_regio_level_profile}{A string.}

\item{type}{A string.}

\item{aikasarja}{A logical.}
}
\description{
Process data for profile
}
