% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alueprofiili.R
\name{create_gt_tbl}
\alias{create_gt_tbl}
\title{Process data for profile}
\usage{
create_gt_tbl(lst_df = lista1_tbl02)
}
\arguments{
\item{lst_df}{A data frame}
}
\description{
Process data for profile
}
