% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indicator.R
\name{create_municipalities_within_region}
\alias{create_municipalities_within_region}
\title{Create list of region neighbors}
\usage{
create_municipalities_within_region(
  input_value_variable = "Nuorisotyöttömyys",
  input_value_regio_level = "Kunnat",
  input_value_region_selected = "Helsinki",
  timeseries = TRUE,
  dat_ts = NULL,
  dat_cs = NULL
)
}
\arguments{
\item{input_value_variable}{A string.}

\item{input_value_regio_level}{A string.}

\item{input_value_region_selected}{A string.}

\item{timeseries}{A logical.}
}
\description{
Create list of region neighbors
}
