% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alueprofiili.R
\name{create_raw_tbl}
\alias{create_raw_tbl}
\title{Process data for profile}
\usage{
create_raw_tbl(dd1 = tabdat_tmp, muuttujanimi = muuttujanimi, varnro = 1)
}
\arguments{
\item{dd1}{A data frame}

\item{muuttujanimi}{A string.}

\item{varnro}{A integer}
}
\description{
Process data for profile
}
