% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_v20211104_aikasarja}
\alias{df_v20211104_aikasarja}
\title{df_v20211104_aikasarja}
\format{
A data frame with 72142 rows and 8 variables:
\describe{
\item{regio_level}{regio_level}
\item{aluekoodi}{aluekoodi}
\item{aluenimi}{aluenimi}
\item{aika}{aika}
\item{variable}{variable}
\item{value}{value}
\item{var_class}{var_class}
\item{all_upper}{all_upper}
}
}
\usage{
df_v20211104_aikasarja
}
\description{
This is internal data for Karttasovellus shiny app
}
\keyword{datasets}
