% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfzip_v20220105}
\alias{dfzip_v20220105}
\title{dfzip_v20220105}
\format{
A data frame with 11976 rows and 5 variables:
\describe{
\item{regio_level}{regio_level}
\item{aluekoodi}{aluekoodi}
\item{aluenimi}{aluenimi}
\item{variable}{variable}
\item{value}{value}
}
}
\usage{
dfzip_v20220105
}
\description{
Cross-sectional zipcode level attribute data
}
\keyword{datasets}
