% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfzip_v20220202_aikasarja}
\alias{dfzip_v20220202_aikasarja}
\title{dfzip_v20220202_aikasarja}
\format{
A data frame with 104790 rows and 8 variables:
\describe{
\item{regio_level}{regio_level}
\item{aluekoodi}{aluekoodi}
\item{aluenimi}{aluenimi}
\item{kuntanro}{kuntanro}
\item{kuntanimi}{kuntanimi}
\item{aika}{aika}
\item{variable}{variable}
\item{value}{value}
}
}
\usage{
dfzip_v20220202_aikasarja
}
\description{
Time-series zipcode level attribute data
}
\keyword{datasets}
