% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\name{document_data}
\alias{document_data}
\title{Document an internal data}
\usage{
document_data(dat, neim, description = "Data is data")
}
\arguments{
\item{dat}{data.frame}

\item{description}{Description of the data}

\item{heim}{string name of the data}
}
\description{
Document an internal data
}
