% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode_functions.R
\name{get_koodit_zip}
\alias{get_koodit_zip}
\title{Process time-series zipcode data}
\usage{
get_koodit_zip(regio_selected = 161, value_regio_level = "Seutukunnat")
}
\arguments{
\item{regio_selected}{A numeric}

\item{value_regio_level}{A string}
}
\description{
Process time-series zipcode data
}
