% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode_functions.R
\name{get_naapurikoodit_zip}
\alias{get_naapurikoodit_zip}
\title{Process time-series zipcode data}
\usage{
get_naapurikoodit_zip(
  region_data = region_data,
  regio_selected = "Veteli Keskus"
)
}
\arguments{
\item{region_data}{A data.frame}

\item{regio_selected}{A string}
}
\description{
Process time-series zipcode data
}
