% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ineq_data}
\alias{ineq_data}
\title{ineq_data}
\format{
A data frame with 17087 rows and 7 variables:
\describe{
\item{aluenimi}{aluenimi}
\item{var_class}{var_class}
\item{variable}{variable}
\item{aika}{aika}
\item{gini}{gini}
\item{regio_level}{regio_level}
\item{aluekoodi}{aluekoodi}
}
}
\usage{
ineq_data
}
\description{
municipality level weighted ginis for all indicators at various regional breakdowns
}
\keyword{datasets}
