% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode_functions.R
\name{map_zipcodes}
\alias{map_zipcodes}
\title{Create zipcode map}
\usage{
map_zipcodes(
  input_value_region_selected = "Veteli Keskus",
  input_value_regio_show_mode = "kaikki tason alueet",
  input_value_variable = "Kokonaislukema",
  leaflet = FALSE
)
}
\arguments{
\item{input_value_region_selected}{A string}

\item{input_value_regio_show_mode}{A string}

\item{input_value_variable}{A string}

\item{leaflet}{A logical}
}
\description{
Create zipcode map
}
