% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_zipcode.R
\name{map_zipcodes_alueprofiili}
\alias{map_zipcodes_alueprofiili}
\title{Create leaflet only zipcode map for alueprofiili}
\usage{
map_zipcodes_alueprofiili(
  input_value_region_selected = 91,
  input_value_regio_level = "Kunnat",
  zipvars = c("Kokonaislukema", "Alimpaan tuloluokkaan kuuluvat taloudet",
    "Alimpaan tuloluokkaan kuuluvat täysi-ikäiset", "Työttömät",
    "Peruskoulutuksen omaavat")
)
}
\arguments{
\item{input_value_region_selected}{A numeric}

\item{input_value_regio_show_mode}{A string}
}
\description{
Create leaflet only zipcode map for alueprofiili
}
