% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{muuttujakuvaukset}
\alias{muuttujakuvaukset}
\title{muuttujakuvaukset}
\format{
A data frame with 32 rows and 4 variables:
\describe{
\item{Muuttujaluokka}{Muuttujaluokka}
\item{Muuttuja}{Muuttuja}
\item{Aluetasot}{Aluetasot}
\item{Kuvaus}{Kuvaus}
}
}
\usage{
muuttujakuvaukset
}
\description{
This is internal data for Karttasovellus shiny app
}
\keyword{datasets}
