% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indicator.R
\name{plot_map}
\alias{plot_map}
\title{Create municipality map}
\usage{
plot_map(
  input_value_regio_level = "Hyvinvointialueet",
  input_value_variable = "Huono-osaisuus yhteensä",
  input_value_region_selected = "Varsinais-Suomen HVA",
  input_value_regio_show_mode = "valittu alue ja sen naapurit",
  leaflet = FALSE
)
}
\arguments{
\item{input_value_regio_level}{A string.}

\item{input_value_variable}{A string.}

\item{input_value_region_selected}{A string.}

\item{input_value_regio_show_mode}{A string.}

\item{leaflet}{A logical.}
}
\description{
Create municipality map
}
