% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indicator.R
\name{plot_rank_bar}
\alias{plot_rank_bar}
\title{Create bar plot}
\usage{
plot_rank_bar(
  input_value_regio_level = "Seutukunnat",
  input_value_variable = "Nuorisotyöttömyys",
  input_value_region_selected = "Helsinki",
  input_value_regio_show_mode = "valitun alueen kunnat"
)
}
\arguments{
\item{input_value_regio_level}{A string.}

\item{input_value_variable}{A string.}

\item{input_value_region_selected}{A string.}

\item{input_value_regio_show_mode}{A string.}
}
\description{
Create bar plot
}
