% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indicator.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Create timeseries plot}
\usage{
plot_timeseries(
  input_value_regio_level = "Kunnat",
  input_value_variable = "Nuorisotyöttömyys",
  input_value_region_selected = "Helsinki",
  input_value_regio_show_mode = "valittu alue ja sen naapurit"
)
}
\arguments{
\item{input_value_regio_level}{A string.}

\item{input_value_variable}{A string.}

\item{input_value_region_selected}{A string.}

\item{input_value_regio_show_mode}{A string.}
}
\description{
Create timeseries plot
}
