% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_zipcode.R
\name{plot_zipcodes_bar}
\alias{plot_zipcodes_bar}
\title{Create zipcode bar chart}
\usage{
plot_zipcodes_bar(
  input_value_region_selected = 5,
  input_value_regio_level = "Hyvinvointialueet",
  input_value_variable = "Kokonaislukema"
)
}
\arguments{
\item{input_value_region_selected}{A numeric}

\item{input_value_regio_level}{A string}

\item{input_value_variable}{A string}
}
\description{
Create zipcode bar chart
}
