% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_zipcode.R
\name{plot_zipcodes_dotplot_alueprofiili}
\alias{plot_zipcodes_dotplot_alueprofiili}
\title{Create zipcode bar chart}
\usage{
plot_zipcodes_dotplot_alueprofiili(
  input_value_region_selected = 5,
  input_value_regio_level = "Hyvinvointialueet",
  zipvars = c("Kokonaislukema", "Alimpaan tuloluokkaan kuuluvat taloudet",
    "Alimpaan tuloluokkaan kuuluvat täysi-ikäiset", "Työttömät",
    "Peruskoulutuksen omaavat")
)
}
\arguments{
\item{input_value_region_selected}{A numeric}

\item{input_value_regio_level}{A string}

\item{zipvars}{A string}
}
\description{
Create zipcode bar chart
}
