% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode_functions.R
\name{plot_zipcodes_line}
\alias{plot_zipcodes_line}
\title{Create zipcode line chart}
\usage{
plot_zipcodes_line(
  input_value_region_selected = "Veteli Keskus",
  input_value_regio_show_mode = "kaikki tason alueet",
  input_value_variable = "Kokonaislukema"
)
}
\arguments{
\item{input_value_region_selected}{A string}

\item{input_value_regio_show_mode}{A string}

\item{input_value_variable}{A string}
}
\description{
Create zipcode line chart
}
