% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_data}
\alias{pop_data}
\title{pop_data}
\format{
A data frame with 4010 rows and 5 variables:
\describe{
\item{regio_level}{regio_level}
\item{aluenimi}{aluenimi}
\item{aika}{aika}
\item{pop}{pop}
\item{aluekoodi}{aluekoodi}
}
}
\usage{
pop_data
}
\description{
Population data for computing population weighted gini coefficients
}
\keyword{datasets}
