% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alueprofiili.R
\name{print_add_line_break2}
\alias{print_add_line_break2}
\title{Add line breaks}
\usage{
print_add_line_break2(
  x = "very many many characters and words and sentences",
  n = 20
)
}
\arguments{
\item{x}{A string.}

\item{n}{A integer}
}
\description{
Add line breaks
}
