% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alueprofiili.R
\name{print_create_tabdat}
\alias{print_create_tabdat}
\title{Create table for print}
\usage{
print_create_tabdat(
  params_region_level = "Kunnat",
  params_region = "Alajärvi",
  params_naapurikoodit = c(91, 924, 92)
)
}
\arguments{
\item{params_region_level}{a string.}

\item{params_region}{a string.}

\item{params_naapurikoodit}{a integer.}
}
\description{
Create table for print
}
