% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alueprofiili.R
\name{print_create_table}
\alias{print_create_table}
\title{Create table for print}
\usage{
print_create_table(
  varclass = "Summamuuttujat",
  valid_vars = "Huono-osaisuus yhteensä",
  tabdat2 = tabdat
)
}
\arguments{
\item{varclass}{A string.}

\item{df}{A data frame.}
}
\description{
Create table for print
}
