% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alueprofiili.R
\name{print_create_timeseries_plot}
\alias{print_create_timeseries_plot}
\title{Create table for print}
\usage{
print_create_timeseries_plot(
  varclass = "Summamuuttujat",
  reglevel = "Kunnat",
  regname = "Alajärvi",
  naapurit = c(91, 92),
  valid_vars = "Huono-osaisuus yhteensä",
  ncol = 2
)
}
\arguments{
\item{varclass}{a string.}

\item{reglevel}{a string.}

\item{ncol}{a integer.}

\item{spatdat}{a sf}

\item{naapurikoodit}{a integer.}
}
\description{
Create table for print
}
