% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indicator.R
\name{process_data}
\alias{process_data}
\title{Process municipality data}
\usage{
process_data(
  input_value_regio_level = "Kunnat",
  input_value_variable = "Nuorisotyöttömyys",
  timeseries = FALSE,
  spatial = TRUE,
  dat_ts = NULL,
  dat_cs = NULL
)
}
\arguments{
\item{input_value_regio_level}{A string.}

\item{input_value_variable}{A string.}

\item{timeseries}{A logical.}

\item{dat_ts}{A data frame. An optional data to speed up things}

\item{dat_cs}{A data frame. An optional data to speed up things}
}
\description{
Process municipality data
}
