% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alueprofiili.R
\name{process_data_profile_doc}
\alias{process_data_profile_doc}
\title{Process data for profile}
\usage{
process_data_profile_doc(input_value_regio_level_profile = "Hyvinvointialueet")
}
\arguments{
\item{input_value_regio_level_profile}{A string.}
}
\description{
Process data for profile
}
