% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode_functions.R
\name{process_zipdata}
\alias{process_zipdata}
\title{Process cross-sectional zipcode data}
\usage{
process_zipdata(varname = "Kokonaislukema")
}
\arguments{
\item{varname}{A string.}
}
\description{
Process cross-sectional zipcode data
}
