% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_zipcode.R
\name{process_zipdata_timeseries}
\alias{process_zipdata_timeseries}
\title{Process time-series zipcode data}
\usage{
process_zipdata_timeseries(varname = "Kokonaislukema")
}
\arguments{
\item{varname}{A string.}
}
\description{
Process time-series zipcode data
}
