% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regio_Kunnat}
\alias{regio_Kunnat}
\title{regio_Kunnat}
\format{
A data frame with 293 rows and 2 variables:
\describe{
\item{aluekoodi}{aluekoodi}
\item{geom}{geom}
}
}
\usage{
regio_Kunnat
}
\description{
This is internal data for Karttasovellus shiny app
}
\keyword{datasets}
