% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regio_Postinumeroalueet}
\alias{regio_Postinumeroalueet}
\title{regio_Postinumeroalueet}
\format{
A data frame with 3027 rows and 4 variables:
\describe{
\item{region_code}{region_code}
\item{region_name}{region_name}
\item{kuntanro}{kuntanro}
\item{geom}{geom}
}
}
\usage{
regio_Postinumeroalueet
}
\description{
Zipcode sf data from 2021 including zipcode and municipality number
}
\keyword{datasets}
