% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regio_Suomi}
\alias{regio_Suomi}
\title{regio_Suomi}
\format{
A data frame with 1 rows and 2 variables:
\describe{
\item{maa}{maa}
\item{geom}{geom}
}
}
\usage{
regio_Suomi
}
\description{
This is internal data for Karttasovellus shiny app
}
\keyword{datasets}
