% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_data}
\alias{region_data}
\title{region_data}
\format{
A data frame with 381 rows and 5 variables:
\describe{
\item{region_code}{region_code}
\item{region_name}{region_name}
\item{geom}{geom}
\item{level}{level}
\item{neigbours}{neigbours}
}
}
\usage{
region_data
}
\description{
This is internal data for Karttasovellus shiny app
}
\keyword{datasets}
