% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_data_zip}
\alias{region_data_zip}
\title{region_data_zip}
\format{
A data frame with 3027 rows and 6 variables:
\describe{
\item{region_code}{region_code}
\item{region_name}{region_name}
\item{kuntanro}{kuntanro}
\item{neigbours}{neigbours}
\item{geom}{geom}
\item{level}{level}
}
}
\usage{
region_data_zip
}
\description{
Zipcode sf with neighbours from 2021
}
\keyword{datasets}
