% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_zipcode.R
\name{table_zipcodes}
\alias{table_zipcodes}
\title{Create zipcode table}
\usage{
table_zipcodes(
  input_value_region_selected = 91,
  input_value_regio_level = "Kunnat"
)
}
\arguments{
\item{input_value_region_selected}{A numeric}

\item{input_value_regio_level}{A string}
}
\description{
Create zipcode table
}
