% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alueprofiili.R
\name{alueprofiiliaikasarja_html}
\alias{alueprofiiliaikasarja_html}
\title{Create leaflet map
Process data for profile}
\usage{
alueprofiiliaikasarja_html(
  input_value_regio_level_profile = "Hyvinvointialueet",
  input_value_region_profile = "Itä-Uudenmaan HVA",
  val_muuttujaluokka = "Summamuuttujat",
  val_muuttuja1 = "Huono-osaisuus yhteensä",
  region_data = NULL,
  tabdat = NULL,
  naapurikoodit = NULL
)
}
\arguments{
\item{input_value_regio_level_profile}{A string.}

\item{input_value_region_profile}{A string.}

\item{val_muuttujaluokka}{A string.}

\item{val_muuttuja}{A string.}
}
\description{
Create leaflet map
Process data for profile
}
