% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfzip_v20220105_aikasarja}
\alias{dfzip_v20220105_aikasarja}
\title{dfzip_v20220105_aikasarja}
\format{
A data frame with 71856 rows and 6 variables:
\describe{
\item{regio_level}{regio_level}
\item{aluekoodi}{aluekoodi}
\item{aluenimi}{aluenimi}
\item{aika}{aika}
\item{variable}{variable}
\item{value}{value}
}
}
\usage{
dfzip_v20220105_aikasarja
}
\description{
Time-series zipcode level attribute data
}
\keyword{datasets}
