% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode_functions.R
\name{get_naapurikoodit_zip}
\alias{get_naapurikoodit_zip}
\title{Process time-series zipcode data}
\usage{
get_naapurikoodit_zip(regio_selected = 5)
}
\arguments{
\item{regio_selected}{A string}

\item{region_data}{A data.frame}
}
\description{
Process time-series zipcode data
}
