% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_zipcode.R
\name{map_zipcodes}
\alias{map_zipcodes}
\title{Create zipcode map}
\usage{
map_zipcodes(
  input_value_region_selected = 72,
  input_value_regio_level = "Kunnat",
  input_value_variable = "Kokonaislukema",
  leaflet = FALSE,
  alueprofiili = FALSE,
  alueprofiili_doc = FALSE,
  basesize = 6,
  plottextsize = 2.5,
  add_caption = TRUE
)
}
\arguments{
\item{input_value_region_selected}{A numeric}

\item{input_value_variable}{A string}

\item{leaflet}{A logical}

\item{alueprofiili}{A logical}

\item{alueprofiili_doc}{A logical}

\item{input_value_regio_show_mode}{A string}
}
\description{
Create zipcode map
}
