% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_zipcode.R
\name{plot_zipcodes_line}
\alias{plot_zipcodes_line}
\title{Create zipcode line chart}
\usage{
plot_zipcodes_line(
  input_value_region_selected = 91,
  input_value_regio_level = "Kunnat",
  input_value_variable = "Kokonaislukema"
)
}
\arguments{
\item{input_value_region_selected}{A numeric}

\item{input_value_variable}{A string}

\item{input_value_regio_show_mode}{A string}
}
\description{
Create zipcode line chart
}
