% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indicator.R
\name{process_data}
\alias{process_data}
\title{Process municipality data}
\usage{
process_data(
  input_value_regio_level = "Kunnat",
  input_value_variable = "Nuorisotyöttömyys",
  timeseries = FALSE
)
}
\arguments{
\item{input_value_regio_level}{A string.}

\item{input_value_variable}{A string.}

\item{timeseries}{A logical.}
}
\description{
Process municipality data
}
